+++ src/opcode/MargValueType.h
@@ -1,9 +1,9 @@
 #ifndef __MARG_VALUE_TYPE_H_
 #define __MARG_VALUE_TYPE_H_
 
-#include <stdlib.h> /* uint64_t */
+#include <stddef.h> /* size_t */
 
 /** @brief Ex nihilo value */
-typedef uint64_t MargValue;
+typedef size_t MargValue;
 
 #endif
